Ajuste a vista de Historial Medico
Se ajusta y agrega la vista para consultar el historial medico de los pacientes (Usol de la Librería de Datatable) y adicional se agregan las imágenes que usa el proyecto.
