+++ .angular/cache/17.3.6/vite/deps/@fortawesome_angular-fontawesome.js
@@ -1,6 +1,7 @@
 import {
   DomSanitizer
-} from "./chunk-G3FTWH2I.js";
+} from "./chunk-XKWNE4EL.js";
+import "./chunk-B35CBZSK.js";
 import {
   Component,
   Directive,
