+++ .angular/cache/17.3.6/vite/deps/_metadata.json
@@ -1,43 +1,52 @@
 {
-  "hash": "507349e3",
-  "configHash": "810f041c",
-  "lockfileHash": "8b7a087a",
-  "browserHash": "af1a2b96",
+  "hash": "481217b1",
+  "configHash": "df2fd3c6",
+  "lockfileHash": "cc917a72",
+  "browserHash": "e86d078b",
   "optimized": {
     "@angular/core": {
       "src": "../../../../../node_modules/@angular/core/fesm2022/core.mjs",
       "file": "@angular_core.js",
-      "fileHash": "f77c4926",
+      "fileHash": "e0b90efb",
       "needsInterop": false
     },
     "@angular/platform-browser": {
       "src": "../../../../../node_modules/@angular/platform-browser/fesm2022/platform-browser.mjs",
       "file": "@angular_platform-browser.js",
-      "fileHash": "16dd8713",
+      "fileHash": "254a818d",
       "needsInterop": false
     },
     "@angular/router": {
       "src": "../../../../../node_modules/@angular/router/fesm2022/router.mjs",
       "file": "@angular_router.js",
-      "fileHash": "9b1b45d9",
+      "fileHash": "88428a1a",
       "needsInterop": false
     },
     "@fortawesome/angular-fontawesome": {
       "src": "../../../../../node_modules/@fortawesome/angular-fontawesome/fesm2022/angular-fontawesome.mjs",
       "file": "@fortawesome_angular-fontawesome.js",
-      "fileHash": "13068c2d",
+      "fileHash": "cf149238",
       "needsInterop": false
     },
     "@fortawesome/free-solid-svg-icons": {
       "src": "../../../../../node_modules/@fortawesome/free-solid-svg-icons/index.mjs",
       "file": "@fortawesome_free-solid-svg-icons.js",
-      "fileHash": "f989c68b",
+      "fileHash": "27b01fe8",
+      "needsInterop": false
+    },
+    "angular-datatables": {
+      "src": "../../../../../node_modules/angular-datatables/fesm2022/angular-datatables.mjs",
+      "file": "angular-datatables.js",
+      "fileHash": "7cfa5d15",
       "needsInterop": false
     }
   },
   "chunks": {
-    "chunk-G3FTWH2I": {
-      "file": "chunk-G3FTWH2I.js"
+    "chunk-XKWNE4EL": {
+      "file": "chunk-XKWNE4EL.js"
+    },
+    "chunk-B35CBZSK": {
+      "file": "chunk-B35CBZSK.js"
     },
     "chunk-AUDQI6TG": {
       "file": "chunk-AUDQI6TG.js"
