+++ package.json
@@ -24,8 +24,13 @@
     "@fortawesome/angular-fontawesome": "^0.14.1",
     "@fortawesome/fontawesome-free": "^6.5.2",
     "@fortawesome/free-solid-svg-icons": "^6.5.2",
+    "@types/datatables.net": "^1.12.0",
+    "angular-datatables": "^17.1.0",
     "bootstrap": "^5.3.3",
+    "datatables.net": "^2.0.5",
+    "datatables.net-dt": "^2.0.5",
     "express": "^4.18.2",
+    "jquery": "^3.7.1",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.14.3"
@@ -36,6 +41,7 @@
     "@angular/compiler-cli": "^17.3.0",
     "@types/express": "^4.17.17",
     "@types/jasmine": "~5.1.0",
+    "@types/jquery": "^3.5.29",
     "@types/node": "^18.18.0",
     "jasmine-core": "~5.1.0",
     "karma": "~6.4.0",
