+++ src/app/components/login/login.component.html
@@ -4,7 +4,7 @@
 
             <div class="header-login">
                 <div class="circle-header"></div>
-                <img src="../../../assets/images/Icon1.png" alt="">
+                <img src="../../../assets/images/images/Icon1.png" alt="">
             </div>
 
             <div class="title-login">
