+++ src/app/dashboard/dashboard/dashboard.component.html
@@ -12,7 +12,7 @@
         <div class="col-xs-12 col-sm-12 col-md-3 col-lg-3">
             <div class="row card-counter history">
                 <div class="img1 col-2">
-                    <img src="../../../assets/images/Icon2.png" alt="" width="70">
+                    <img src="../../../assets/images/images/Icon2.png" alt="" width="70">
                 </div>
                 <div class="col-9">
                     <span class="count-name">Historial Clinico</span>
@@ -23,7 +23,7 @@
         <div class="col-xs-12 col-sm-12 col-md-3 col-lg-3">
             <div class="card-counter warning">
                 <div class="img2 col-2">
-                    <img src="../../../assets/images/Icon1.png" alt="" width="70">
+                    <img src="../../../assets/images/images/Icon1.png" alt="" width="70">
                 </div>
                 <div class="col-9">
                     <span class="count-name">Medicamentos</span>
@@ -34,7 +34,7 @@
         <div class="col-xs-12 col-sm-12 col-md-3 col-lg-3">
             <div class="row card-counter info">
                 <div class="img3 col-2">
-                    <img src="../../../assets/images/Icon4.png" alt="" width="70">
+                    <img src="../../../assets/images/images/Icon4.png" alt="" width="70">
                 </div>
                 <div class="col-9">
                     <span class="count-name">Citas Medicas</span>
