+++ src/app/dashboard/historial-medico/historial-medico.component.css
@@ -13,3 +13,85 @@
     height: 100vh;
 }
 
+.nav {
+  margin: 50px;
+}
+
+.nav .nav-link{
+  padding: 15px;
+  border-radius: 0;
+}
+
+.nav-link{
+  color: #396172;
+}
+
+.nav-link.active{
+  border-bottom: 3px solid #000;
+  background-color: #396172;
+  color: #fff;
+}
+
+.tab-pane {
+  margin-top: 50px;
+  width: 100vh !important;
+  height: auto;
+  padding: 10px;
+  background-color: #fff;
+}
+
+.tab-pane.active{
+  border-left: 5px solid #396172;
+}
+
+.resumenClinico{
+  width: 90%;
+  height: auto;
+  background-color: #fff;
+  margin-top: 50px;
+  text-align: center;
+}
+
+.titleHistorialMedico{
+  background-color: #396172;
+  color: #fff;
+  text-align: center;
+  border-radius: 15px;
+  margin-bottom: 10px;
+}
+
+.opciones{
+  margin: 15px;
+}
+
+.btn-opciones{
+  border-radius: 10px;
+  color: #fff;
+  width: 26px;
+  height: 28px;
+  margin: 3px;
+}
+
+.alergias {
+  display: flex;
+  justify-content: center;
+}
+
+.alergias .alert{
+  background-color: #ffebeb;
+  color: red;
+  width: 75%;
+  height: auto;
+}
+
+.botonesHM{
+  background-color: #396172;
+  color: #fff;
+  margin: 5px;
+}
+
+.botonesHM:hover{
+  background: rgb(54,93,111);
+  background: linear-gradient(90deg, rgba(54,93,111,1) 0%, rgba(87,132,148,1) 50%, rgba(142,162,158,1) 100%);
+  transition: all 0.5s ease;
+}
