+++ src/app/dashboard/historial-medico/historial-medico.component.html
@@ -1,25 +1,101 @@
 <div class="main">
+
+
     <div class="row">
 
-        <section class="col-8">
+        <section class="col-9">
+
             <div class="d-flex align-items-start">
                 <div class="nav flex-column nav-pills me-3" id="v-pills-tab" role="tablist" aria-orientation="vertical">
-                    <button class="nav-link active" id="historial" data-bs-toggle="pill" data-bs-target="#v-pills-home" type="button" role="tab" aria-controls="v-pills-home" aria-selected="true">Historial</button>
-                    <button class="nav-link" id="notas" data-bs-toggle="pill" data-bs-target="#v-pills-profile" type="button" role="tab" aria-controls="v-pills-profile" aria-selected="false">Notas</button>
-                    <button class="nav-link" id="examenFisico" data-bs-toggle="pill" data-bs-target="#v-pills-messages" type="button" role="tab" aria-controls="v-pills-messages" aria-selected="false">Examen Fisico</button>
-                    <button class="nav-link" id="diagnostico" data-bs-toggle="pill" data-bs-target="#v-pills-settings" type="button" role="tab" aria-controls="v-pills-settings" aria-selected="false">Diagnostico y Tratamiento</button>
-                    <button class="nav-link" id="medicamentos" data-bs-toggle="pill" data-bs-target="#v-pills-settings" type="button" role="tab" aria-controls="v-pills-settings" aria-selected="false">Medicamentos</button>
+                    <button class="nav-link active" id="historial" data-bs-toggle="pill" data-bs-target="#p-historial" type="button" role="tab" aria-controls="p-historial" aria-selected="true">Historial</button>
+                    <button class="nav-link" id="diagnostico" data-bs-toggle="pill" data-bs-target="#p-diagnostico" type="button" role="tab" aria-controls="p-diagnostico" aria-selected="false">Diagnostico y Tratamiento</button>
+                    <button class="nav-link" id="medicamentos" data-bs-toggle="pill" data-bs-target="#p-medicamentos" type="button" role="tab" aria-controls="p-medicamentos" aria-selected="false">Medicamentos</button>
                 </div>
                 <div class="tab-content" id="v-pills-tabContent">
-                    <div class="tab-pane fade show active" id="v-pills-home" role="tabpanel" aria-labelledby="v-pills-home-tab">...</div>
-                    <div class="tab-pane fade" id="v-pills-profile" role="tabpanel" aria-labelledby="v-pills-profile-tab">...</div>
-                    <div class="tab-pane fade" id="v-pills-messages" role="tabpanel" aria-labelledby="v-pills-messages-tab">...</div>
-                    <div class="tab-pane fade" id="v-pills-settings" role="tabpanel" aria-labelledby="v-pills-settings-tab">...</div>
+                    <div class="tab-pane fade show active" id="p-historial" role="tabpanel" aria-labelledby="historial">
+                        
+                        <div class="col-12 titleHistorialMedico">
+                            <p>Datos Paciente</p>
+                        </div>
+
+                        <div class="row d-flex justify-content-around">
+                            <div class="col-3 d-flex justify-content-center">
+                                <fa-icon class="icons-nav" [icon]="faUserCircle" style="font-size: 7rem;"></fa-icon>
+                            </div>
+                            <div class="col-4 p-3" style="border: 1px dashed #ddd;">
+                                <b>Paciente: </b><p>Juan Alberto Perez </p>
+                                <b>Fecha Nacimiento: </b><p>Juan Perez</p>
+                                <b>Dirección: </b><p>Juan Perez</p>
+                            </div>
+                            <div class="col-4 p-3" style="border: 1px dashed #ddd;">
+                                <b>Teléfono: </b><p>Juan Perez</p>
+                                <b>Genero: </b><p>Juan Perez</p>
+                                <b>Email: </b><p>Juan Perez</p>
+                            </div>
+                        </div>
+
+                        <div class="col-12 titleHistorialMedico">
+                            <p>Historial Clinico</p>
+                        </div>
+
+                        <div class="col-12 d-flex justify-content-end">
+                            <button class="btn botonesHM col-2">Registrar</button>
+                            <button class="btn botonesHM col-2">Editar</button>
+                        </div>
+
+                        <div class="infoHistorial p-3">
+                            <table datatable [dtOptions]="dtOptions" class="table hover"></table>
+                        </div>
+                    </div>
+                    
+                    <div class="tab-pane fade" id="p-diagnostico" role="tabpanel" aria-labelledby="diagnostico">
+
+                        <div class="col-12 titleHistorialMedico">
+                            <p>Diagnostico de Enfermedades</p>
+                        </div>
+
+                        <div class="infoHistorial p-3">
+                            <table datatable [dtOptions]="dtOptionsDT" class="table hover" style="width: 100%;"></table>
+                        </div>
+
+                    </div>
+
+                    <div class="tab-pane fade" id="p-medicamentos" role="tabpanel" aria-labelledby="medicamentos">
+
+                        <div class="col-12 titleHistorialMedico">
+                            <p>Medicamentos</p>
+                        </div>
+
+                        <div class="infoHistorial p-3">
+                            <table datatable [dtOptions]="dtOptionsM" class="table hover" style="width: 100%;"></table>
+                        </div>
+
+                    </div>
                 </div>
             </div>
+
         </section>  
 
-        <section class="col-4"></section>
+        <section class="col-3">
+            <div class="resumenClinico p-2">
+                <h4 class="mt-4"><b>Resumen Clinico</b></h4>
+
+                <section class="opciones">
+                    <button class="btn btn-opciones btn-info"><fa-icon class="icons-nav" [icon]="faDownload"></fa-icon></button>
+                    <button class="btn btn-opciones btn-success"><fa-icon class="icons-nav" [icon]="faEye"></fa-icon></button>
+                </section>
+
+                <section class="alergias">
+                    <div class="alert p-2">
+                        <b><fa-icon class="icons-nav" [icon]="faExclamationTriangle"></fa-icon> Alergias a Medicamentos</b>
+                        <p> - Penincilina</p>
+                        <b>Otras Alergias: </b>
+                        <p> - Nueces</p>
+                    </div>
+                </section>
+
+            </div>
+        </section>
 
     </div>
 </div>
